Move SerializableErrorInfo and associates to dagster-shared
