+++ python_modules/dagster/dagster/_cli/definitions.py
@@ -3,6 +3,7 @@ import os
 import sys
 
 import click
+from dagster_shared.error import remove_system_frames_from_error
 
 from dagster import __version__ as dagster_version
 from dagster._cli.utils import assert_no_remaining_opts, get_possibly_temporary_instance_for_cli
@@ -11,7 +12,7 @@ from dagster._cli.workspace.cli_target import (
     get_workspace_from_cli_opts,
     workspace_options,
 )
-from dagster._utils.error import remove_system_frames_from_error, unwrap_user_code_error
+from dagster._utils.error import unwrap_user_code_error
 from dagster._utils.log import configure_loggers
